+++ src/lib/sound/Sound.cpp
@@ -3,7 +3,7 @@
 Sound::Sound() {
 }
 
-void Sound::appendAudio(double audio[], int len) {
+void Sound::appendAudio(float audio[], int len) {
     for(int i=0; i<len; i++) {
         buffer.push_back(audio[i]);
     }
